sql-proxy-adapter/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
PyYAML==6.0.1
sqlglot==19.0.0